Move utilities to SILOptimizer/Utils/Differentiaton.cpp
