+++ include/swift/SILOptimizer/Utils/Differentiation.h
@@ -49,10 +49,31 @@ namespace autodiff {
 /// This is being used to print short debug messages within the AD pass.
 raw_ostream &getADDebugStream();
 
+/// Returns true if this is an `ApplyInst` with `array.uninitialized_intrinsic`
+/// semantics.
+bool isArrayLiteralIntrinsic(ApplyInst *ai);
+
 /// If the given value `v` corresponds to an `ApplyInst` with
 /// `array.uninitialized_intrinsic` semantics, returns the corresponding
 /// `ApplyInst`. Otherwise, returns `nullptr`.
 ApplyInst *getAllocateUninitializedArrayIntrinsic(SILValue v);
+
+/// Given an element address from an `array.uninitialized_intrinsic` `apply`
+/// instruction, returns the `apply` instruction. The element address is either
+/// a `pointer_to_address` or `index_addr` instruction to the `RawPointer`
+/// result of the instrinsic:
+///
+///     %result = apply %array.uninitialized_intrinsic : $(Array<T>, RawPointer)
+///     (%array, %ptr) = destructure_tuple %result
+///     %elt0 = pointer_to_address %ptr to $*T       // element address
+///     %index_1 = integer_literal $Builtin.Word, 1
+///     %elt1 = index_addr %elt0, %index_1           // element address
+///     ...
+ApplyInst *getAllocateUninitializedArrayIntrinsicElementAddress(SILValue v);
+
+/// Given a value, finds its single `destructure_tuple` user if the value is
+/// tuple-typed and such a user exists.
+DestructureTupleInst *getSingleDestructureTupleUser(SILValue value);
 
 /// Given an `apply` instruction, apply the given callback to each of its
 /// direct results. If the `apply` instruction has a single `destructure_tuple`
