+++ lib/SILOptimizer/Mandatory/Differentiation.cpp
@@ -78,108 +78,11 @@ static llvm::cl::opt<bool> SkipFoldingDifferentiableFunctionExtraction(
 // Helpers
 //===----------------------------------------------------------------------===//
 
-
 /// Given a dumpable value, dumps it to `llvm::dbgs()`.
 template <typename T> static inline void debugDump(T &v) {
   LLVM_DEBUG(llvm::dbgs() << "\n==== BEGIN DEBUG DUMP ====\n"
                           << v << "\n==== END DEBUG DUMP ====\n");
 }
-
-namespace swift {
-namespace autodiff {
-
-raw_ostream &getADDebugStream() { return llvm::dbgs() << "[AD] "; }
-
-static bool isArrayLiteralIntrinsic(ApplyInst *ai) {
-  return ai->hasSemantics("array.uninitialized_intrinsic");
-}
-
-ApplyInst *getAllocateUninitializedArrayIntrinsic(SILValue v) {
-  if (auto *ai = dyn_cast<ApplyInst>(v))
-    if (isArrayLiteralIntrinsic(ai))
-      return ai;
-  return nullptr;
-}
-
-/// Given an element address from an `array.uninitialized_intrinsic` `apply`
-/// instruction, returns the `apply` instruction. The element address is either
-/// a `pointer_to_address` or `index_addr` instruction to the `RawPointer`
-/// result of the instrinsic:
-///
-///     %result = apply %array.uninitialized_intrinsic : $(Array<T>, RawPointer)
-///     (%array, %ptr) = destructure_tuple %result
-///     %elt0 = pointer_to_address %ptr to $*T       // element address
-///     %index_1 = integer_literal $Builtin.Word, 1
-///     %elt1 = index_addr %elt0, %index_1           // element address
-///     ...
-static ApplyInst *
-getAllocateUninitializedArrayIntrinsicElementAddress(SILValue v) {
-  // Find the `pointer_to_address` result, peering through `index_addr`.
-  auto *ptai = dyn_cast<PointerToAddressInst>(v);
-  if (auto *iai = dyn_cast<IndexAddrInst>(v))
-    ptai = dyn_cast<PointerToAddressInst>(iai->getOperand(0));
-  if (!ptai)
-    return nullptr;
-  // Return the `array.uninitialized_intrinsic` application, if it exists.
-  if (auto *dti = dyn_cast<DestructureTupleInst>(
-          ptai->getOperand()->getDefiningInstruction())) {
-    if (auto *ai = getAllocateUninitializedArrayIntrinsic(dti->getOperand()))
-      return ai;
-  }
-  return nullptr;
-}
-
-/// Given a value, finds its single `destructure_tuple` user if the value is
-/// tuple-typed and such a user exists.
-static DestructureTupleInst *getSingleDestructureTupleUser(SILValue value) {
-  bool foundDestructureTupleUser = false;
-  if (!value->getType().is<TupleType>())
-    return nullptr;
-  DestructureTupleInst *result = nullptr;
-  for (auto *use : value->getUses()) {
-    if (auto *dti = dyn_cast<DestructureTupleInst>(use->getUser())) {
-      assert(!foundDestructureTupleUser &&
-             "There should only be one `destructure_tuple` user of a tuple");
-      foundDestructureTupleUser = true;
-      result = dti;
-    }
-  }
-  return result;
-}
-
-void forEachApplyDirectResult(
-    ApplyInst *ai, llvm::function_ref<void(SILValue)> resultCallback) {
-  if (!ai->getType().is<TupleType>()) {
-    resultCallback(ai);
-    return;
-  }
-  if (auto *dti = getSingleDestructureTupleUser(ai))
-    for (auto result : dti->getResults())
-      resultCallback(result);
-}
-
-void collectAllFormalResultsInTypeOrder(SILFunction &function,
-                                        SmallVectorImpl<SILValue> &results) {
-  SILFunctionConventions convs(function.getLoweredFunctionType(),
-                               function.getModule());
-  auto indResults = function.getIndirectResults();
-  auto *retInst = cast<ReturnInst>(function.findReturnBB()->getTerminator());
-  auto retVal = retInst->getOperand();
-  SmallVector<SILValue, 8> dirResults;
-  if (auto *tupleInst =
-          dyn_cast_or_null<TupleInst>(retVal->getDefiningInstruction()))
-    dirResults.append(tupleInst->getElements().begin(),
-                      tupleInst->getElements().end());
-  else
-    dirResults.push_back(retVal);
-  unsigned indResIdx = 0, dirResIdx = 0;
-  for (auto &resInfo : convs.getResults())
-    results.push_back(resInfo.isFormalDirect() ? dirResults[dirResIdx++]
-                                               : indResults[indResIdx++]);
-}
-
-} // end namespace autodiff
-} // end namespace swift
 
 /// Given a function, gathers all of its direct results in an order defined by
 /// its result type. Note that "formal results" refer to result values in the
