+++ lib/SILOptimizer/Utils/CMakeLists.txt
@@ -8,6 +8,7 @@ silopt_register_sources(
   ConstExpr.cpp
   # SWIFT_ENABLE_TENSORFLOW
   DerivativeLookup.cpp
+  Differentiation.cpp
   # SWIFT_ENABLE_TENSORFLOW END
   Devirtualize.cpp
   Existential.cpp
